refactor: organise unit tests
