+++ SeaPublicWebsite.UnitTests/BusinessLogic/Models/ResetUnusedFieldsTests.cs
@@ -3,7 +3,7 @@ using NUnit.Framework;
 using SeaPublicWebsite.BusinessLogic.Models;
 using SeaPublicWebsite.BusinessLogic.Models.Enums;
 
-namespace Tests;
+namespace Tests.BusinessLogic.Models;
 
 public class ResetUnusedFieldsTests
 {
